+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: anjose-d <anjose-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/03 00:11:28 by anjose-d          #+#    #+#             */
-/*   Updated: 2021/08/03 00:11:28 by anjose-d         ###   ########.fr       */
+/*   Updated: 2021/08/09 22:45:37 by anjose-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (!dst || !src)
