+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: anjose-d <anjose-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/03 00:11:32 by anjose-d          #+#    #+#             */
-/*   Updated: 2021/08/03 00:11:33 by anjose-d         ###   ########.fr       */
+/*   Updated: 2021/08/09 22:45:58 by anjose-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	unsigned int	count;
+	size_t	count;
 
 	count = 0;
 	while (str[count])
